Merged, Inheritance stuff, Worked on implementing carryable ingredients
Merged. Improved the OOD structure and got started on the ingredients object. Got started on trying to make that object follow the player that is holding it (without a "grab" method, this has to happen in the editor in runtime or at compile time). A key problem preventing that from working is that the finalMove vector is returning (0,0,0). It's not a collision problem preventing movement, because the bowl of wheat can be "held" without collision. I may be approaching this entirely wrong, but if we can get that movement vector, we can move the object to the correct position as the player moves around.