Ingredients will follow the holding player
Ingredients will follow either the duck or the human, and they will pivot to be in front of the direction you're moving. This is done with an OOD implementation, so we can easily add more ingredients with the same behavior.

There is still no input implementation, so you need to run the game, then go into the editor and manually switch the "isHeld" and "holder" attributes, then click back to the game in order to actually hold the bowl of wheat.